[Android] Remove frames from error report. Refactoring.
